requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-cors-headers==3.14.0
django-phonenumber-field==7.1.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack==1.0.5
packaging==23.0
paho-mqtt==1.6.1
phonenumbers==8.13.18
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.4
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sentry-sdk==1.19.1
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
